Changed value back to a DOUBLE now its the final commit to lazy to make delete lowercase.

+++ ZotighP3/ZotighP3.cpp
@@ -17,7 +17,7 @@ int main()
 	string usersName;
 	string usersChoice;
 	string choice;
-	int planetChoice;
+	double planetChoice;
 	double percentageIlluminated;
 
 	// vector of stings for the names
